backend/requirements.txt
django
djangorestframework
python-dotenv
requests
django-cors-headers
